requirements.txt
matplotlib==3.1.1
numpy==1.17.0
pandas==0.25.0
Pillow==6.1.0
scikit-learn==0.21.3
scipy==1.3.0
six==1.12.0
torch==1.1.0
torchvision==0.3.0
opencv-python==4.1.0.25